Update cclip.hpp file and improve code compatibility
The cclip.hpp file was changed significantly to improve compatibility with older versions of C++. The code no longer has a requirement for C++17 or later versions. The inclusion of the <cstring> library and the use of strcpy instead of strcpy_s also contributes to making the code more adaptable.